0x04 task 12 commit 5

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -11,7 +11,7 @@ void print_number(int n)
 {
 	unsigned int m; /* power of 10 */
 	int c; /* boolean check */
-	long num; /* convert int to long */
+	unsigned int num; /* convert int to long */
 
 	num = n;
 	/* negatives */
